+++ UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/Collaboration/Runtime/UMI3DCollaborationClientUserTracking.cs
@@ -15,8 +15,10 @@ limitations under the License.
 */
 
 using System.Collections;
+using System.Collections.Generic;
 using System.Linq;
 using umi3d.cdk.userCapture;
+using umi3d.common.userCapture;
 using UnityEngine;
 
 namespace umi3d.cdk.collaboration
@@ -73,6 +75,101 @@ namespace umi3d.cdk.collaboration
             yield return new WaitUntil(() => UMI3DCollaborationClientServer.Connected());
 
             base.DispatchCamera();
+        }
+
+        /// <summary>
+        /// Collection of emotes' coroutine related to playing for each user.
+        /// </summary>
+        private Dictionary<ulong, Coroutine> emoteCoroutineDict = new Dictionary<ulong, Coroutine>();
+
+        /// <summary>
+        /// Starts an emote on another user's avatar in the scene.
+        /// </summary>
+        /// <param name="emoteId">Emote to start UMI3D Id.</param>
+        /// <param name="userId">Id of the user to start the emote for.</param>
+        /// Don't use this for your own avatar's emotes.
+        public void PlayEmoteOnOtherAvatar(ulong emoteId, ulong userId)
+        {
+            var otherUserAvatar = embodimentDict[userId];
+            var animators = otherUserAvatar.GetComponentsInChildren<Animator>();
+            var emoteAnimator = animators.Where(animator => animator.runtimeAnimatorController != null).FirstOrDefault();
+
+            if (emoteAnimator == null || emoteConfig == null)
+                return;
+            var emoteToPlay = emoteConfig.emotes.Find(x => x.id == emoteId);
+            if (emoteCoroutineDict.ContainsKey(userId) && emoteCoroutineDict[userId] != null) //an Emote is playing, need to interrupt it
+            {
+                StopCoroutine(emoteCoroutineDict[userId]);
+                emoteAnimator.enabled = false;
+                emoteAnimator.Update(0);
+            }
+
+            var coroutine = StartCoroutine(PlayEmote(emoteAnimator, emoteToPlay));
+            if (emoteCoroutineDict.ContainsKey(userId))
+                emoteCoroutineDict[userId] = coroutine;
+            else
+                emoteCoroutineDict.Add(userId, coroutine);
+        }
+
+        /// <summary>
+        /// Plays an emote on an animator.
+        /// </summary>
+        /// <param name="animator">Animator to play the emote on.</param>
+        /// <param name="emote">Emote to play.</param>
+        /// <returns></returns>
+        protected IEnumerator PlayEmote(Animator animator, UMI3DEmoteDto emote)
+        {
+            animator.enabled = true;
+            animator.Play(emote.stateName);
+            animator.Update(0);
+            yield return new WaitWhile(() =>
+            {
+                if (animator == null) return false;  // heppens when a user leaves the scene when playing an emote
+                return animator.GetCurrentAnimatorStateInfo(0).normalizedTime < 1;
+            });
+            if (animator == null) // heppens when a user leaves the scene when playing an emote
+                yield break;
+            animator.enabled = false;
+        }
+
+        /// <summary>
+        /// Stops an emote on another user's avatar in the scene.
+        /// </summary>
+        /// <param name="emoteId">Emote to stop UMI3D Id.</param>
+        /// <param name="userId">Id of the user to stop the emote for.</param>
+        /// Don't use this for your own avatar's emotes.
+        public void StopEmoteOnOtherAvatar(ulong emoteId, ulong userId)
+        {
+            if (emoteConfig == null) //no emote support in the scene
+                return;
+
+            if (emoteCoroutineDict.ContainsKey(userId)
+                && emoteCoroutineDict[userId] != null)
+            {
+                StopCoroutine(emoteCoroutineDict[userId]);
+                emoteCoroutineDict[userId] = null;
+
+                if (UMI3DClientUserTracking.Instance.embodimentDict.TryGetValue(userId, out UserAvatar otherUserAvatar))
+                {
+                    if (otherUserAvatar == null) //the embodiment system lost the avatar
+                        return;
+
+                    var animators = otherUserAvatar.GetComponentsInChildren<Animator>();
+                    if (animators == null) //no animator to desactive found on the avatar
+                        return;
+
+                    var emoteAnimator = animators.Where(animator => animator.runtimeAnimatorController != null).FirstOrDefault();
+                    if (emoteAnimator == null) //no animator to desactive found on the avatar
+                        return;
+
+                    var emoteToStop = emoteConfig.emotes.Find(x => x.id == emoteId);
+                    if (emoteAnimator == null) //the emote to stop doesn't exist
+                        throw new Umi3dException("The emote to stop does not exist in emote configuration file.");
+
+                    emoteAnimator.Update(0);
+                    emoteAnimator.enabled = false;
+                }
+            }
         }
     }
 }
